refactor(search): rename widget and fix styles
