+++ lib/app/pages/search/search_page.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:solar_system_umpontoseis/app/shared/utils/data.dart';
 import 'package:solar_system_umpontoseis/app/shared/widgets/search_input.dart';
 
-import 'widgets/header.dart';
 import 'widgets/planet_suggestions.dart';
 import 'widgets/result_planet_card.dart';
+import 'widgets/search_header.dart';
 
 class SearchPage extends StatelessWidget {
   const SearchPage({Key key}) : super(key: key);
